requirements.txt
Django==5.0.7
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.4
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
pytest==8.3.2
pytest-django==4.9.0
factory_boy==3.3.0
Faker==26.0.0